% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{biplot}
\alias{biplot}
\title{Plot com ggplot2}
\usage{
biplot(x, y, upper, lower)
}
\arguments{
\item{x}{series a serem plotadas}

\item{y}{opcional serie ou series a serem plotadas juntas para comparacao}

\item{upper}{limite superior para regiao sombreada}

\item{lower}{limite inferior para regiao sombreada}
}
\value{
ggplot das series
}
\description{
Plot com ggplot2
}

