% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabelas.R
\name{cointeg}
\alias{cointeg}
\title{Teste de cointegracao de Johansen}
\usage{
cointeg(x, ecdet, K, season = NULL)
}
\arguments{
\item{x}{Data matrix to be investigated for cointegration}

\item{ecdet}{Character, "none" for no intercept in 
cointegration, "const" for constant term in cointegration 
and "trend" for trend variable in cointegration.}

\item{K}{The lag order of the series (levels) in the VAR}

\item{season}{If seasonal dummies should be included, the
data frequency must be set accordingly, i.e "4" for 
quarterly data.}
}
\value{
A list with the following elements:
  \item{\code{eigen}}{estatistica do teste do lambda-max}
  \item{\code{trace}}{estatistica do teste do traco}
  
  @import urca
  
  @export
}
\description{
Conducts the Johansen procedure on a given data set. 
The "trace" or "eigen" statistics are reported. 
A wrapper for the \link[urca]{ca.jo} function in the urca package.
}

