% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{ic}
\alias{ic}
\title{Intervalo de confianca}
\usage{
ic(x, dp, nc = 0.95)
}
\arguments{
\item{x}{estimativa pontual}

\item{dp}{desvio padrao da estimativa pontual}

\item{nc}{nivel de confianca do intervalo, entre 0 e 1}
}
\value{
uma lista com 
\item{\code{upper}}{limite superior do intervalo}
\item{\code{lower}}{limite inferior do intervalo}
}
\description{
Calcula o intervalo de confianca para uma estimativa pontual
}
