% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucmodel.R
\name{ic.ucmodel}
\alias{ic.ucmodel}
\title{Intervalo de confianca para ucmodel}
\usage{
ic.ucmodel(ucm, state, type = "smooth", ic = 0.95)
}
\arguments{
\item{ucm}{um modelo estimado por ucmodel}

\item{state}{o nome do estado estimado para o qual se deve
construir o intervalo de confianca}

\item{type}{tipo do estado usado, suavizado (padrao) ou filtrado}

\item{ic}{o nivel de confianca do intervalo, entre 0 e 1}
}
\value{
uma lista com 
\item{\code{upper}}{limite superior do intervalo}
\item{\code{lower}}{limite inferior do intervalo}
}
\description{
Calcula o intervalo de confianca de Bonferroni para um
componente estimado por ucmodel
}
