% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabelas.R
\name{n2tab}
\alias{n2tab}
\title{Transforma numero e texto formatado}
\usage{
n2tab(x)
}
\arguments{
\item{x}{um objeto com valores numericos}
}
\value{
o mesmo objeto que foi fornecido com numeros 
transformados em texto formatado
}
\description{
Transforma numero e texto formatado para serem mostrados na
tabela, o numero arredondado para 3 casas decimais e separado
por virgula
}

