% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucmodel.R
\name{pars.ucmodel}
\alias{pars.ucmodel}
\title{Extrair os paramentros de objetos ucmodel}
\usage{
pars.ucmodel(x)
}
\arguments{
\item{x}{um objeto "ucmodel"}
}
\value{
uma lista com ar (parametros autoregressivos), sigma
(vaiancias) e rho (covariancias)
}
\description{
Extrair os paramentros do modelo de componentes nao observaveis
estimado por ucmodel
}

