% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucmodel.R
\name{SSMarima}
\alias{SSMarima}
\title{Create a State Space Model Object Arima of Class SSModel}
\usage{
SSMarima(ar = NULL, ma = NULL, d = 0, Q, stationary = TRUE, index,
  n = 1, ynames)
}
\arguments{
\item{ar}{a list of numeric vector containing the autoregressive coeffients.}

\item{ma}{nao usado (sempre NULL).}

\item{d}{a degree of differencing (nao usado, sempre 0)}

\item{Q}{a p x p covariance matrix of the disturbances 
(or in the time varying case p x p x n array), where 
$p$ = length(index)}

\item{stationary}{logical value indicating whether a stationarity of the 
arima part is assumed. Defaults to TRUE.}

\item{index}{A vector indicating for which series the corresponding 
components are constructed.}

\item{n}{Length of the series, only used internally for dimensionality check.}

\item{ynames}{names of the times series, used internally.}
}
\value{
Object of class SSModel (ver \link[KFAS]{SSModel})
}
\description{
Adaptadacao da funcao \link[KFAS]{SSMarima} do package KFAS 
para permitir parametros regressivos diferentes entre as series
}

