% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabelas.R
\name{tab.cointeg}
\alias{tab.cointeg}
\title{Tabela para o teste de cointegracao}
\usage{
tab.cointeg(x)
}
\arguments{
\item{x}{um objeto mts com series temporais multivariadas}
}
\value{
tabela com as estatisticas dos testes
}
\description{
Efetua o teste de cointegracao de Johansen para series e 
fornece uma tabela com as estatisticas
}
