% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucmodel.R
\name{tab.ucmodel}
\alias{tab.ucmodel}
\title{Tabela com as estimativas do ucmodel}
\usage{
tab.ucmodel(ucm)
}
\arguments{
\item{ucm}{objeto ucmodel com modelo estimado por ucmodel}
}
\value{
uma tabela com as estimativas
}
\description{
Tabela com as estimativas do ucmodel
}
