% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucmodel.R
\name{ucmodel}
\alias{ucmodel}
\title{Estima o modelo de componentes nao observaveis com tendencia
e ciclo por maxima verossimelhanca}
\usage{
ucmodel(x, l, init, corre = TRUE)
}
\arguments{
\item{x}{um objeto mts com a serie temporia multivariada}

\item{l}{numero de defagens do modelo AR}

\item{init}{valores iniciais dos parametros a serem estimados}

\item{corre}{logical, se TRUE (padrao) modelo com componentes 
correlacionado, se FALSE modelo com componentes restrito se 
correlacao}
}
\value{
uma lista com os seguinte objetos fit (\link[KFAS]{fitSSM}) e out (\link[KFAS]{KFS})
}
\description{
Estima o modelo de componentes nao observaveis com tendencia
e ciclo por maxima verossimelhanca usando o pacote "KFAS"
}

