% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucmodel.R
\name{ucplot}
\alias{ucplot}
\title{Plota ucmodel com ggplot2}
\usage{
ucplot(ucm, state, type = "smooth", ic = 0.95)
}
\arguments{
\item{ucm}{um modelo estimado por ucmodel}

\item{state}{o nome do estado estimado para o qual se deve
plotar o grafico}

\item{type}{tipo do estado usado, suavizado (padrao) ou filtrado}

\item{ic}{o nivel de confianca do intervalo, entre 0 e 1}
}
\value{
ggplot das series
}
\description{
Plota ucmodel com ggplot2
}

